backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==39.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==5.1.0
Markdown==3.4.1
nodeenv==1.9.1
numpy==1.25.1
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
pycparser==2.21
PyJWT==1.7.1
pytesseract==0.3.10
pytz==2022.6
sqlparse==0.4.3
XlsxWriter==3.1.2
zipp==3.11.0
